relocated one test to test_stnode.py
