+++ tests/test_stnode.py
@@ -2,9 +2,11 @@ import os
 from contextlib import nullcontext
 
 import asdf
+from asdf.exceptions import ValidationError
 import astropy.units as u
 import pytest
 
+from roman_datamodels import maker_utils as utils
 from roman_datamodels import datamodels, maker_utils, stnode, validate
 from roman_datamodels.testing import assert_node_equal, assert_node_is_copy, wraps_hashable
 
@@ -179,6 +181,26 @@ def test_set_pattern_properties():
     mdl.phot_table.F062.photmjsr = None
     mdl.phot_table.F062.uncertainty = None
     mdl.phot_table.F062.pixelareasr = None
+
+
+# Test that a currently undefined attribute can be assigned using dot notation
+# so long as the attribute is defined in the coresponding schema.
+def test_node_new_attribute_assignment():
+    exp = stnode.Exposure()
+    with pytest.raises(AttributeError):
+        exp.bozo = 0
+    exp.ngroups = 5
+    assert exp.ngroups == 5
+    # Test patternProperties attribute case
+    photmod = utils.mk_wfi_img_photom()
+    phottab = photmod.phot_table
+    newphottab = {'F062': phottab['F062']}
+    photmod.phot_table = newphottab
+    photmod.phot_table.F213 = phottab['F213']
+    with pytest.raises(AttributeError):
+        photmod.phot_table.F214 = phottab['F213']
+    with pytest.raises(ValidationError):
+        photmod.phot_table.F106 = 0
 
 
 VALIDATION_CASES = ("true", "yes", "1", "True", "Yes", "TrUe", "YeS", "foo", "Bar", "BaZ")
